refactoring generate ict/pop
